♻️ move clients in dedicated modules
