+++ backend/common/infrastructure/src/github/mod.rs
@@ -1,8 +1,4 @@
-use std::{
-	collections::HashMap,
-	fmt::Debug,
-	sync::{Arc, Mutex},
-};
+use std::{collections::HashMap, fmt::Debug};
 
 use anyhow::anyhow;
 use async_trait::async_trait;
@@ -10,9 +6,12 @@ use octocrab::{
 	models::{issues::Issue, pulls::PullRequest, repos::Content, Repository, User},
 	FromResponse, Octocrab, OctocrabBuilder,
 };
-use olog::{debug, tracing::instrument};
+use olog::tracing::instrument;
 use reqwest::Url;
 use serde::Deserialize;
+
+mod clients;
+pub use clients::{RoundRobinClient, SingleClient};
 
 mod error;
 pub use error::Error;
@@ -26,78 +25,6 @@ pub struct Config {
 	base_url: String,
 	personal_access_tokens: String,
 	headers: HashMap<String, String>,
-}
-
-pub struct RoundRobinClient {
-	octocrab_clients: Vec<Octocrab>,
-	next_octocrab_clients_index: Arc<Mutex<usize>>,
-}
-
-impl RoundRobinClient {
-	pub fn new(config: &Config) -> anyhow::Result<Self> {
-		let personal_access_tokens: Vec<&str> = config
-			.personal_access_tokens
-			.split(',')
-			.map(|token| token.trim())
-			.filter(|token| !token.is_empty())
-			.collect();
-
-		if personal_access_tokens.is_empty() {
-			return Err(anyhow!(
-				"No Github personal_access_token was provided in configuration"
-			));
-		}
-
-		let octocrab_clients: anyhow::Result<Vec<Octocrab>> = personal_access_tokens
-			.iter()
-			.map(|personal_access_token| {
-				Ok(Octocrab::builder()
-					.base_url(&config.base_url)?
-					.personal_token(personal_access_token.to_string())
-					.add_headers(&config.headers)?
-					.build()?)
-			})
-			.collect();
-
-		debug!(
-			"Github API client setup with {} personal access tokens",
-			personal_access_tokens.len()
-		);
-
-		Ok(Self {
-			octocrab_clients: octocrab_clients?,
-			next_octocrab_clients_index: Arc::new(Mutex::new(0)),
-		})
-	}
-}
-
-impl OctocrabProxy for RoundRobinClient {
-	fn octocrab(&self) -> &Octocrab {
-		let mut index = self.next_octocrab_clients_index.lock().unwrap();
-		let next_octocrab = &self.octocrab_clients[*index];
-		*index = (*index + 1) % self.octocrab_clients.len();
-		next_octocrab
-	}
-}
-
-pub struct SingleClient(Octocrab);
-
-impl SingleClient {
-	pub fn new(config: &Config, personal_access_token: String) -> anyhow::Result<Self> {
-		Ok(Self(
-			Octocrab::builder()
-				.base_url(&config.base_url)?
-				.personal_token(personal_access_token)
-				.add_headers(&config.headers)?
-				.build()?,
-		))
-	}
-}
-
-impl OctocrabProxy for SingleClient {
-	fn octocrab(&self) -> &Octocrab {
-		&self.0
-	}
 }
 
 #[async_trait]
@@ -253,68 +180,5 @@ impl AddHeaders for OctocrabBuilder {
 			self = self.add_header(key.parse()?, value.clone());
 		}
 		Ok(self)
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use rstest::rstest;
-
-	use super::*;
-
-	#[rstest]
-	#[case("http://plop.fr/github/", Some("https://api.github.com/repos/ning-rain/evens/contributors".parse().unwrap()), Some("http://plop.fr/github/repos/ning-rain/evens/contributors".parse().unwrap()))]
-	#[case("http://plop.fr/github", Some("https://api.github.com/repos/ning-rain/evens/contributors".parse().unwrap()), Some("http://plop.fr/github/repos/ning-rain/evens/contributors".parse().unwrap()))]
-	#[case("http://plop.fr/github/", Some("https://api.github.com".parse().unwrap()), Some("http://plop.fr/github/".parse().unwrap()))]
-	#[case("http://plop.fr/github", Some("https://api.github.com".parse().unwrap()), Some("http://plop.fr/github/".parse().unwrap()))]
-	#[case("http://plop.fr/github/", None, None)]
-	fn fix_github_host(
-		#[case] base_url: &str,
-		#[case] url: Option<reqwest::Url>,
-		#[case] expected_url: Option<reqwest::Url>,
-	) {
-		let client = RoundRobinClient::new(&Config {
-			base_url: base_url.to_string(),
-			personal_access_tokens: "token".to_string(),
-			headers: HashMap::new(),
-		})
-		.unwrap();
-
-		let result_url = client.fix_github_host(&url).unwrap();
-		assert_eq!(result_url, expected_url);
-	}
-
-	#[rstest]
-	#[case("only_one_token".to_string(), 1)]
-	#[case("token_a,token_b".to_string(), 2)]
-	#[case("token_a, token_b, token_c,,,token_d, token_e,  ,   , token_f".to_string(), 6)]
-	fn personal_access_tokens_config(#[case] personal_access_tokens: String, #[case] count: usize) {
-		let client = RoundRobinClient::new(&Config {
-			base_url: "http://plop.fr/github/".to_string(),
-			personal_access_tokens,
-			headers: HashMap::new(),
-		})
-		.unwrap();
-
-		assert_eq!(client.octocrab_clients.len(), count);
-
-		// check the round robin doesn't go out of bounds
-		for _ in 0..(count * 3) {
-			client.octocrab();
-		}
-	}
-
-	#[rstest]
-	#[case("".to_string())]
-	#[case(",".to_string())]
-	#[case(", ,,, ,".to_string())]
-	fn invalid_personal_access_tokens_config(#[case] personal_access_tokens: String) {
-		let result = RoundRobinClient::new(&Config {
-			base_url: "http://plop.fr/github/".to_string(),
-			personal_access_tokens,
-			headers: HashMap::new(),
-		});
-
-		assert!(result.is_err());
 	}
 }
